Moved some vector functions around
